Create KeyboardToLEDs.cpp
This C++ program read an input from a connected keyboard, and used that input to adjust a counter, which was then reflected by the state of the LEDs on a connected FPGA board. When the up key is pressed, the counter is incremented by 1; when the down key is pressed, the counter is decremented by 1; when the left key is pressed, the counter is shifted left by one bit; when the right key is pressed, the counter is shifted right by one bit; finally, when the R key is pressed, the counter resets to 0.